Refactors home.js by pulling out class declarations into their own file and importing what is needed.
